% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_variants_to_dir.R
\name{chunk_variants_to_dir}
\alias{chunk_variants_to_dir}
\title{Title}
\usage{
chunk_variants_to_dir(
  chunk_name,
  chunk_name_suffix = "_variants",
  file_name = NULL,
  dir = "R/",
  replace1,
  replacements1,
  replace2 = NULL,
  replacements2 = NULL,
  replace3 = NULL,
  replacements3 = NULL,
  replace4 = NULL,
  replacements4 = NULL
)
}
\arguments{
\item{chunk_name}{}

\item{chunk_name_suffix}{}

\item{file_name}{}

\item{dir}{}

\item{replace1}{}

\item{replacements1}{}

\item{replace2}{}

\item{replacements2}{}

\item{replace3}{}

\item{replacements3}{}

\item{replace4}{}

\item{replacements4}{}
}
\description{
Title
}
