% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitcodegetlive.R
\name{return_chunk_code}
\alias{return_chunk_code}
\title{Title}
\usage{
return_chunk_code(chunk_name)
}
\arguments{
\item{chunk_name}{a character string with the name of the chunk of interest}
}
\value{
a vector of the code contained in the referenced chunk
}
\description{
Title
}
